# (failed?) Attempt to inline Primitive.cexpr!