--- a/tool/mk_builtin_loader.rb
+++ b/tool/mk_builtin_loader.rb
@@ -311,7 +311,7 @@ def mk_builtin_header file
           f.puts %'        fprintf(f, "    const VALUE *argv = GET_EP() - lnum - VM_ENV_DATA_SIZE + 1 + %ld;\\n", index);'
           f.puts %'    }'
         end
-        f.puts %'    fprintf(f, "    func f = (func)/* #{cfunc_name} */%p;\\n", (const void *)#{cfunc_name});'
+        f.puts %'    fprintf(f, "    func f = (func)%"PRIdPTR"; /* == #{cfunc_name} */\\n", (intptr_t)#{cfunc_name});'
         f.puts %'    fprintf(f, "    val = f(ec, GET_SELF()#{argv});\\n");'
       end
       f.puts %'}'