Noticed that because `Primitive.cexpr!` is a purely compile-time construct, we can eliminate runtime `lookup_builtin_invoker()` circus and directly embed the expressions themselves into JITed codes.

<details><summary>This works, and for instance generated inline C for `Integer#abs` would become:</summary>

```C
static inline VALUE
_mjit0_inlined_2(rb_execution_context_t *ec, rb_control_frame_t *reg_cfp, const VALUE orig_self, const rb_iseq_t *original_iseq)
{
    const VALUE *orig_pc = reg_cfp->pc;
    VALUE *orig_sp = reg_cfp->sp;
    VALUE stack[1];
    static const VALUE *const original_body_iseq = (VALUE *)0x559e523c18e0;

label_0: /* opt_invokebuiltin_delegate_leave */
{
    VALUE val;
    MAYBE_UNUSED(VALUE) self = GET_SELF();
{
    val = rb_int_abs(self);
}
    stack[0] = val;
}
    return stack[0];

(... snip ...)
```

</details>

<details><summary>However IT SLOWS DOWN on AWS EC2 c6g.2xlarge:</summary>

```
$ benchmark-driver -v --repeat-count=11 --output=all --rbenv 'master::master@git;ours::tmp;master --jit::master@git --jit;ours --jit::tmp --jit' --filter=abs ~/data/src/github.com/shy
ouhei/ruby/benchmark/mjit_integer.yml
master: ruby 2.8.0dev (2020-07-09T22:01:10Z master ba81bc24e6) [aarch64-linux]
ours: ruby 2.8.0dev (2020-07-10T02:49:50Z origin/compile-bui.. 2d5c59e21e) [aarch64-linux]
master --jit: ruby 2.8.0dev (2020-07-09T22:01:10Z master ba81bc24e6) +JIT [aarch64-linux]
ours --jit: ruby 2.8.0dev (2020-07-10T02:49:50Z origin/compile-bui.. 2d5c59e21e) +JIT [aarch64-linux]
Calculating -------------------------------------
                                   master                  ours          master --jit            ours --jit
        mjit_abs(-1)   18489311.253129158    18982457.009280208    30128929.258191332    29354327.567977905 i/s
                       19080117.484867666    19237473.860359915    30399283.531448383    29402037.452557422
                       19202604.364737410    19295726.280229516    30715431.769021541    29460415.679664187
                       19429883.302648768    19403658.373607460    30748517.479827501    29532670.372629836
                       19569566.973248113    19439899.586268935    30812276.965962008    29593321.386331223
                       19601113.757824801    19593888.245184083    30873217.203203499    29998978.497288637
                       19656988.673114762    19606039.929396592    30879054.025119048    30158585.142605729
                       19706913.073762778    19715792.173321426    30923353.643916830    30385699.086196225
                       19912591.172063068    19847260.327253465    31360081.653118704    30825604.444944806
                       19965823.561417315    20097527.836062040    31516313.017349727    31053188.264961582
                       20079694.389420200    20233826.909296039    31837865.136915416    31475418.785401657
```

</details>