@shyouhei 2020-07-10
JFYI `perf diff` (of master versus ours in above benchmark) is like:

![Screenshot from 2020-07-10 13-03-05](https://user-images.githubusercontent.com/15377/87115351-eb3d1100-c2ad-11ea-806f-e962acbefe66.png)