@k0kubun 2020-07-10
Cool :) It was certainly a plan to truly inline expressions, but I've been too lazy to think about the actual implementation. You're quick.

> However IT SLOWS DOWN on AWS EC2 c6g.2xlarge:

Interesting... do you have any insight about this? Though, as I believe this is the right direction, please feel free to merge it. We could revisit fixing the slowdown afterwards (possibly reverting this as a last resort).