@shyouhei 2020-07-10
> > However IT SLOWS DOWN on AWS EC2 c6g.2xlarge:
> 
> Interesting... do you have any insight about this?

Not much :worried:  I suspect it could be something related to instruction cache, but `perf stat` shows virtually nothing on this cloud.