@k0kubun 2020-07-10
```
$ benchmark-driver -v --rbenv 'before --jit;after --jit' benchmark/mjit_integer.yml benchmark/mjit_kernel.yml benchmark/mjit_struct_aref.yml --repeat-count=4
before --jit: ruby 2.8.0dev (2020-07-09T22:01:10Z master ba81bc24e6) +JIT [x86_64-linux]
after --jit: ruby 2.8.0dev (2020-07-10T05:54:01Z compile-builtin-di.. c2b208b7b0) +JIT [x86_64-linux]
Calculating -------------------------------------
                         before --jit  after --jit
            mjit_abs(-1)      71.144M      70.823M i/s -     40.000M times in 0.562242s 0.564786s
    mjit_bit_length(100)      68.883M      38.557M i/s -     40.000M times in 0.580698s 1.037419s
           mjit_even?(2)      71.194M      71.415M i/s -     40.000M times in 0.561843s 0.560103s
        mjit_integer?(0)      81.214M      81.408M i/s -     40.000M times in 0.492527s 0.491353s
      mjit_magnitude(-1)      71.011M      70.210M i/s -     40.000M times in 0.563290s 0.569721s
            mjit_odd?(1)      71.300M      39.342M i/s -     40.000M times in 0.561009s 1.016722s
             mjit_ord(1)      82.123M      81.649M i/s -     40.000M times in 0.487075s 0.489901s
            mjit_to_i(1)      82.091M      82.262M i/s -     40.000M times in 0.487264s 0.486250s
          mjit_to_int(1)      81.817M      81.614M i/s -     40.000M times in 0.488899s 0.490111s
           mjit_zero?(0)      71.959M      74.370M i/s -     40.000M times in 0.555868s 0.537852s
        mjit_class(self)      69.605M      69.992M i/s -     40.000M times in 0.574668s 0.571492s
           mjit_class(1)      69.223M      69.004M i/s -     40.000M times in 0.577842s 0.579675s
       mjit_frozen?(str)      72.689M      68.965M i/s -     40.000M times in 0.550292s 0.580006s
      mjit_frozen?(fstr)      71.965M      69.154M i/s -     40.000M times in 0.555823s 0.578423s
mjit_struct_aref(struct)      58.662M      59.923M i/s -     40.000M times in 0.681875s 0.667526s
```

"undefined symbol: rb_int_bit_length" and "undefined symbol: rb_int_odd_p" are happening.

I suspected a link failure for MinGW test_jit failure on `#unpack`, but it doesn't have 'inline' attribute :thinking: