@k0kubun 2020-07-10
For what it's worth, if we eliminate any call insn from Integer paths by inlining the `rb_int_abs` definition like https://github.com/k0kubun/ruby/commit/a410aba827b42084f73da58fe22ec034e58e87d1 (caution: this is a PoC-quality patch), the degraded benchmark could rather be faster:

```
$ benchmark-driver -v --rbenv 'before --jit;after --jit' benchmark/mjit_integer.yml --filter abs --repeat-count=4
before --jit: ruby 2.8.0dev (2020-07-09T22:01:10Z master ba81bc24e6) +JIT [x86_64-linux]
after --jit: ruby 2.8.0dev (2020-07-10T05:54:01Z compile-builtin-di.. c2b208b7b0) +JIT [x86_64-linux]
Calculating -------------------------------------
                     before --jit  after --jit
        mjit_abs(-1)      70.901M      72.529M i/s -     40.000M times in 0.564169s 0.551505s

Comparison:
                     mjit_abs(-1)
         after --jit:  72528865.1 i/s
        before --jit:  70900686.8 i/s - 1.02x  slower
```