@k0kubun 2020-07-10
It might be an option to have two different attributes for ones just eliminating a frame push and ones also inlining a definition so that we inline a definition only when it contributes to performance.